SparseUtils

SparseUtils

Utilities for sparse matrices

The source repository is https://github.com/jlapeyre/SparseUtils.jl.

Contents

SparseUtils

Contents

Index

Functions

Index

Base.nnz

SparseUtils.c_to_julia_index

SparseUtils.c_to_julia_index!

SparseUtils.sparse_stats

SparseUtils.sparsity

SparseUtils.transpose_concrete

Functions

SparseUtils.sparsity — Function.

sparsity(sp::SparseMatrixCSC)

Compute the sparisty of sp. This assumes that there are no stored values equal to zero.

source

SparseUtils.sparse_stats — Function.

sparse_stats(sp::SparseMatrixCSC)

Print some characteristics of sp.

source

Base.nnz — Function.

nnz(sp::SparseMatrixCSC, colnum::Integer)

Return the number of structurally non-zero elements in the colnumth column of sp.

source

SparseUtils.c_to_julia_index! — Function.

c_to_julia_index!(colptr0, rowval0, nzval)::SparseArrays.SparseMatrixCSC

Convert a sparse matrix with zero-based indices to a SparseMatrixCSC. colptr0 and rowval0 are altered in place. nzval is not copied.

source

SparseUtils.c_to_julia_index — Function.

c_to_julia_index(colptr0, rowval0, nzval)::SparseArrays.SparseMatrixCSC

Convert a sparse matrix with zero-based indices to a SparseMatrixCSC. colptr0 and rowval0 are copied and altered. nzval is not copied.

source

SparseUtils.transpose_concrete — Function.

transpose_concrete(M::SparseMatrixCSC)

Return the concrete transpose of the sparse matrix M. That is, return a new sparse matrix, ratherM wrapped in Transpose.

source