SparseUtils

SparseUtils

Utilities for sparse matrices

The source repository is https://github.com/jlapeyre/SparseUtils.jl.

This package implements the type SparseMatrixCOO, several functions for type SparseMatrixCSC, and a few functions for AbstractSparseMatrix. Submodules NZVal and IJV contain the generic algorithms, which are wrapped by the sparse matrix types.

Several functions are not yet documented.

Contents

SparseUtils

Contents

Index

SparseMatrixCOO

SparseMatrixCSC

AbstractSparseMatrix

NZVal

IJV

Index

SparseUtils.IJV

SparseUtils.NZVal

SparseUtils.SparseMatrixCOO

Base.issorted

SparseArrays.nnz

SparseArrays.sparse

SparseArrays.spzeros

SparseUtils.IJV.findindex

SparseUtils.IJV.getindex

SparseUtils.IJV.issorted

SparseUtils.c_to_julia_index

SparseUtils.c_to_julia_index!

SparseUtils.density

SparseUtils.nnzcounts

SparseUtils.prunecols

SparseUtils.prunerows

SparseUtils.summarystats

SparseMatrixCOO

SparseUtils.SparseMatrixCOO — Type.

SparseMatrixCOO{Tv,Ti<:Integer} <: AbstractSparseMatrix{Tv,Ti}

Matrix type for storing sparse matrices in the Coordinate list form (COO) format.

source

SparseArrays.sparse — Function.

sparse(I, J, V, m, n; sparsetype=SparseMatrixCOO)

Create a sparse matrix of type SparseMatrixCOO

source

Base.issorted — Function.

issorted(S::SparseMatrixCOO)

Return true if the data structures in S storing non-zero indices are canonically sorted. If the data structures are not sorted, then functions taking an SparseMatrixCOO argument will, in general, give incorrect results.

source

SparseArrays.spzeros — Function.

spzeros([type,] m,n; sparsetype=SparseMatrixCOO)

Create a sparse matrix of size mxn. If sparsetype is ommited, a matrix of type SparseMatrixCSC will be created.

source

SparseUtils.prunecols — Function.

prunecols(S::SparseMatrixCOO, min_entries; renumber=true)

Remove all columns from S that have fewer than min_entries non-zero entries. If renumber is true, renumber the remaining columns consecutively starting from 1.

source

SparseUtils.prunerows — Function.

prunerows(S::SparseMatrixCOO, min_entries; renumber=true)

Remove all rows from S that have fewer than min_entries non-zero entries. If renumber is true, renumber the remaining rows consecutively starting from 1.

source

SparseMatrixCSC

SparseArrays.nnz — Method.

nnz(sp::SparseMatrixCSC, colnum::Integer)

Return the number of structurally non-zero elements in the colnumth column of sp.

source

SparseUtils.nnzcounts — Function.

nnzcounts(sp::SparseMatrixCSC; rev=true, byvalue=true)

Return a sorted countmap of the number of structural non-zeros in the columns of sp. The keys controlling the sorting are passed to StatsBase.sort.

Examples

In this example, there are 132781 columns with one non-zero element, 32096 columns with two, etc.

julia> conns = nnzcounts(sp);

julia> collect(Iterators.take(conns,3))
3-element Array{Pair{Int64,Int64},1}:
 1 => 132781
 2 => 32096
 3 => 17478

source

SparseUtils.c_to_julia_index! — Function.

c_to_julia_index!(colptr, rowval, nzval)::SparseArrays.SparseMatrixCSC

Convert a sparse matrix with zero-based indices to a SparseMatrixCSC. colptr and rowval are altered in place. nzval is neither copied nor altered.

source

SparseUtils.c_to_julia_index — Function.

c_to_julia_index(colptr, rowval, nzval)::SparseArrays.SparseMatrixCSC

Convert a sparse matrix with zero-based indices to a SparseMatrixCSC. colptr and rowval are not altered, but rather copied. nzval is neither copied, nor altered.

source

<!– SparseUtils.transpose –>

AbstractSparseMatrix

SparseUtils.density — Function.

density(sp::AbstractSparseMatrix)

Compute the density of sp, the fraction of elements that are non-zero This assumes that there are no stored values equal to zero.

source

SparseUtils.summarystats — Function.

summarystats(sp::SparseMatrixCSC)

Print some statistics for sp. The line "number of non-zeros" counts structural non-zeros.

source

NZVal

SparseUtils.NZVal — Module.

module SparseUtils.NZVal

For sparse matrices, algorithms that depend only on nzval, m, and n.

This code is generic, for types in Core.

The symbols in NZVal should not be exported, not even for convenience. They conflict with functions in Base.

source

IJV

SparseUtils.IJV — Module.

module IJV

Algorithms for sparse matrices stored in COO form.

IJV does not depend on any types not in Core. Do not import any symbol from IJV into another module. The methods in IJV are incompatible with functions in Base with the same (unqualified) name. Instead, use IJV like this

#Example

Base.getindex(S::MySparseMatrixCOO, i::Integer, j::Integer) = IJV.getindex(S.I, S.J, S.V, i, j)

source

SparseUtils.IJV.issorted — Function.

issorted(I::AbstractVector, J::AbstractVector)

Return true if the COO indices I and J are sorted canonically.

source

SparseUtils.IJV.findindex — Function.

findindex(I::AbstractArray, J::AbstractArray, i::Integer, j::Integer)

Return the common linear index into I and J (and V) representing coordinates (i, j), or nothing if there is no value stored for (i, j).

source

SparseUtils.IJV.getindex — Function.

getindex(I, J, V, i::Integer, j::Integer; zeroel=zero(eltype(V)))

Return the element at coordinates (i, j) in the matrix represented by I, J, V, if it exists, and zeroel otherwise.

source